requirements.txt
joblib==1.0.1
numpy==1.21.2
pandas==1.3.2
Pillow==8.3.1
python-dateutil==2.8.2
pytz==2021.1
scikit-learn==0.24.2
scipy==1.7.1
six==1.16.0
sklearn==0.0
threadpoolctl==2.2.0
xlwt==1.3.0
